updates to hydrate function and read operation for ifw
